permisos para came, correcciones visuales, descargar credenciales y formatos desde la vista

+++ src/AppBundle/Controller/Came/SolicitudController.php
@@ -14,19 +14,25 @@ use AppBundle\Service\SolicitudManagerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 
 class SolicitudController extends DIEControllerController
 {
     /**
      * @Route("/solicitud", methods={"GET"}, name="solicitud.index")
+     * @Security("has_role('ROLE_CONSULTAR_SOLICITUDES')")
      */
     public function indexAction(Request $request)
     {
         $perPage = $request->query->get('perPage', 10);
         $page = $request->query->get('page', 1);
+        $delegacion = $this->getUserDelegacionId($request->query->get('delegacion'));
+        if(is_null($delegacion)){
+            throw $this->createAccessDeniedException();
+        }
         $solicitudes = $this->getDoctrine()
             ->getRepository(Solicitud::class)
-            ->getAllSolicitudesByDelegacion(null/*simulado*/, $perPage, $page, $request->query->all());
+            ->getAllSolicitudesByDelegacion($delegacion, $perPage, $page, $request->query->all());
         return $this->render('came/solicitud/index.html.twig', [
             'solicitudes' => $this->get('serializer')->normalize(
                 $solicitudes['data'],
@@ -51,14 +57,19 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/api/solicitud", methods={"GET"}, name="solicitud.index.json")
+     * @Security("has_role('ROLE_CONSULTAR_SOLICITUDES')")
      */
     public function indexApiAction(Request $request)
     {
         $perPage = $request->query->get('perPage', 10);
         $page = $request->query->get('page', 1);
+        $delegacion = $this->getUserDelegacionId($request->query->get('delegacion'));
+        if(is_null($delegacion)){
+            throw $this->createAccessDeniedException();
+        }
         $solicitudes = $this->getDoctrine()
             ->getRepository(Solicitud::class)
-            ->getAllSolicitudesByDelegacion(null/*simulado*/, $perPage, $page, $request->query->all());
+            ->getAllSolicitudesByDelegacion($delegacion, $perPage, $page, $request->query->all());
         return $this->jsonResponse([
             'object' => $this->get('serializer')->normalize(
                 $solicitudes['data'],
@@ -83,28 +94,35 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/solicitud/create", methods={"GET"}, name="solicitud.create")
+     * @Security("has_role('ROLE_AGREGAR_SOLICITUD')")
      */
     public function createAction(Request $request)
     {
         $form = $this->createForm(SolicitudType::class);
+        $this->getUser();
+        $delegacion = $this->getUserDelegacionId($request->query->get('delegacion'));
+        if(is_null($delegacion)){
+            throw $this->createAccessDeniedException();
+        }
         $instituciones = $this->getDoctrine()
             ->getRepository(Institucion::class)
-            ->findAllPrivate();
+            ->findAllPrivate($delegacion);
         $unidades = $this->getDoctrine()
             ->getRepository(Unidad::class)
-            ->getAllUnidadesByDelegacion();
+            ->getAllUnidadesByDelegacion($delegacion);
         return $this->render('came/solicitud/create.html.twig', [
             'form' => $form->createView(),
             'instituciones' => $this->get('serializer')->normalize($instituciones, 'json',
                 ['attributes' => ['id', 'nombre', 'rfc', 'direccion', 'telefono', 'correo', 'sitioWeb', 'fax', 'representante']]),
             'unidades' => $this->get('serializer')->normalize($unidades, 'json',
-                ['attributtes' => ['id', 'nombre']])
+                ['attributes' => ['id', 'nombre']])
         ]);
     }
 
 
     /**
      * @Route("/api/solicitud", methods={"POST"}, name="solicitud.store")
+     * @Security("has_role('ROLE_AGREGAR_SOLICITUD')")
      * @param Request $request
      * @param SolicitudManagerInterface $solicitudManager
      */
@@ -116,9 +134,15 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/solicitud/{id}/edit", methods={"GET"}, name="solicitud.edit")
+     * @param Request $request
+     * @Security("has_role('ROLE_EDITAR_SOLICITUD')")
      */
-    public function editAction($id)
+    public function editAction(Request $request, $id)
     {
+        $delegacion = $this->getUserDelegacionId($request->query->get('delegacion'));
+        if(is_null($delegacion)){
+            throw $this->createAccessDeniedException();
+        }
         $solicitud = $this->getDoctrine()
             ->getRepository(Solicitud::class)
             ->find($id);
@@ -128,12 +152,15 @@ class SolicitudController extends DIEControllerController
                 'Not found for id ' . $id
             );
         }
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
         $instituciones = $this->getDoctrine()
             ->getRepository(Institucion::class)
-            ->findAllPrivate();
+            ->findAllPrivate($delegacion);
         $unidades = $this->getDoctrine()
             ->getRepository(Unidad::class)
-            ->getAllUnidadesByDelegacion();
+            ->getAllUnidadesByDelegacion($delegacion);
         $form = $this->createForm(SolicitudType::class);
         return $this->render('came/solicitud/edit.html.twig', [
             'form' => $form->createView(),
@@ -150,12 +177,13 @@ class SolicitudController extends DIEControllerController
                         'cicloAcademico' => ['id', 'nombre'], 'vigencia', 'label']]
                 ]]),
             'unidades' => $this->get('serializer')->normalize($unidades, 'json',
-                ['attributtes' => ['id', 'nombre']]),
+                ['attributes' => ['id', 'nombre']]),
         ]);
     }
 
     /**
      * @Route("/api/solicitud/{id}", methods={"PUT"}, name="solicitud.update")
+     * @Security("has_role('ROLE_EDITAR_SOLICITUD')")
      * @param Request $request
      * @param SolicitudManagerInterface $solicitudManager
      */
@@ -169,6 +197,9 @@ class SolicitudController extends DIEControllerController
                 'Not found for id ' . $id
             );
         }
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
         $form = $this->createForm(SolicitudType::class);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -180,6 +211,7 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/solicitud/{id}", methods={"GET"}, name="solicitud.show")
+     * @Security("has_role('ROLE_DETALLE_SOLICITUD')")
      */
     public function showAction($id)
     {
@@ -193,12 +225,13 @@ class SolicitudController extends DIEControllerController
             );
         }
 
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
+
         $convenios = $this->getDoctrine()
             ->getRepository(Convenio::class)
             ->getAllBySolicitud($solicitud->getId());
-        $pagosCamposClinicos = $this->getDoctrine()
-            ->getRepository(Pago::class)
-            ->getPagosCampoClinicosBySolicitud($solicitud->getId());
 
         return $this->render('came/solicitud/show.html.twig', [
             'solicitud' => $this->get('serializer')->normalize(
@@ -212,7 +245,8 @@ class SolicitudController extends DIEControllerController
                         'id', 'vigencia', 'vigenciaFormatted','label', 'carrera' => ['id', 'nombre',
                             'nivelAcademico' => ['id', 'nombre']], 'numero'],
                     'lugaresSolicitados', 'lugaresAutorizados', 'horario', 'unidad' => ['id', 'nombre'],
-                    'fechaInicial', 'fechaFinal', 'referenciaBancaria', 'fechaInicialFormatted', 'fechaFinalFormatted'],
+                    'fechaInicial', 'fechaFinal', 'referenciaBancaria', 'fechaInicialFormatted', 'fechaFinalFormatted',
+                    'estatus' => ['id', 'nombre']],
                 'pago' => ['id', 'comprobantePago', 'fechaPago', 'fechaPagoFormatted', 'referenciaBancaria', 'factura' => ['fechaFacturacion', 'id', 'fechaFacturacionFormatted']],
                 'pagos' => ['id', 'comprobantePago', 'fechaPago', 'fechaPagoFormatted','referenciaBancaria', 'factura' => ['fechaFacturacion', 'id', 'fechaFacturacionFormatted']]]
             ]),
@@ -238,6 +272,9 @@ class SolicitudController extends DIEControllerController
                 'Not found for id ' . $id
             );
         }
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($solicitud);
         $response = new JsonResponse(['status' => true, "message" => "Solicitud Eliminada con éxito"]);
@@ -246,6 +283,7 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/api/solicitud/terminar/{id}", methods={"POST"}, name="solicitud.terminar")
+     * @Security("has_role('ROLE_TERMINAR_SOLICITUDES')")
      * @param Request $request
      * @param SolicitudManagerInterface $solicitudManager
      * @param $id
@@ -261,13 +299,16 @@ class SolicitudController extends DIEControllerController
                 'Not found for id ' . $id
             );
         }
-
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
         $solicitudManager->finalizar($solicitud);
         return $this->jsonResponse(['status' => true]);
     }
 
     /**
      * @Route("/api/solicitud/validar_montos/{id}", methods={"POST"}, name="solicitud.store_validar_montos")
+     * @Security("has_role('ROLE_VALIDACION_MONTOS')")
      * @param Request $request
      * @param SolicitudManagerInterface $solicitudManager
      * @param $id
@@ -283,7 +324,9 @@ class SolicitudController extends DIEControllerController
                 'Not found for id ' . $id
             );
         }
-
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
+        }
         $form = $this->createForm(ValidaSolicitudType::class, $solicitud);
         $form->get('montos_pagos')->setData($solicitud->getMontosCarreras());
         $form->handleRequest($request);
@@ -298,6 +341,7 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/solicitud/{id}/validar_montos", methods={"GET"}, name="solicitud.validar_montos")
+     * @Security("has_role('ROLE_VALIDACION_MONTOS')")
      * @param Request $request
      * @param $id
      */
@@ -311,6 +355,10 @@ class SolicitudController extends DIEControllerController
             throw $this->createNotFoundException(
                 'Not found for id ' . $id
             );
+        }
+
+        if(!$this->validarSolicitudDelegacion($solicitud)){
+            throw $this->createAccessDeniedException();
         }
 
         $form = $this->createForm(ValidaSolicitudType::class);
@@ -333,6 +381,7 @@ class SolicitudController extends DIEControllerController
 
     /**
      * @Route("/solicitud/{solicitud_id}/oficio", methods={"GET"}, name="came.solicitud.oficio_montos")
+     * @Security("has_role('ROLE_DESCARGAR_OFICIO_MONTOS')")
      * @param $solicitud_id
      * @return mixed
      */
@@ -342,11 +391,13 @@ class SolicitudController extends DIEControllerController
             ->getRepository(Solicitud::class)
             ->find($solicitud_id);
 
+
         if (!$solicitud) {
             throw $this->createNotFoundException(
                 'Not found for id ' . $solicitud_id
             );
         }
+
         $downloadHandler = $this->get('vich_uploader.download_handler');
         return $downloadHandler->downloadObject($solicitud, 'urlArchivoFile');
     }
